refactor(visualize): move map and web map tools into visualize folder
