+++ pyproject.toml
@@ -43,8 +43,6 @@ dependencies = [
     "libpysal>=4.13.0",
     "esda>=2.7.0",
     "spreg==1.8.3",
-    "folium>=0.15.0",
-    "pydeck>=0.9.0",
 ]
 
 [project.optional-dependencies]
@@ -80,6 +78,11 @@ land-cover = [
     "requests>=2.31",
 ]
 
+visualize = [
+    "folium>=0.15.0",
+    "pydeck>=0.9.0",
+]
+
 all = [
     "pygadm>=0.5.0",
     "cdsapi==0.7.6",
@@ -90,6 +93,8 @@ all = [
     "xarray>=2023.1.0",
     "stackstac>=0.5.0",
     "requests>=2.31",
+    "folium>=0.15.0",
+    "pydeck>=0.9.0",
 ]
 
 [project.scripts]
