+++ src/gis_mcp/__init__.py
@@ -8,5 +8,3 @@ from .rasterio_functions import *
 from .pyproj_functions import *
 from .pysal_functions import * 
 from .save_tool import *
-from .map_tool import *
-from .web_map_tool import *
